chore(journal): rename StorageException to JournalException
